Am modificat metoda deleteItem() astfel încât să verifice dacă obiectul căutat există sau nu în database. Dacă acesta nu există, atunci vom returna un ResponseEntity cu HttpStatus.NOT_FOUND. Altfel, vom șterge acel obiect prin metoda deleteById() și vom returna un ResponseEntity cu HttpStatus.NO_CONTENT.

+++ src/main/java/com/siemens/internship/ItemController.java
@@ -55,11 +55,14 @@ public class ItemController {
         }
     }
 
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteItem(@PathVariable Long id) {
-        itemService.deleteById(id);
-        return new ResponseEntity<>(HttpStatus.CONFLICT);
-    }
+        @DeleteMapping("/{id}")
+        public ResponseEntity<Void> deleteItem(@PathVariable Long id) {
+            if (itemService.findById(id).isEmpty()) {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
+            itemService.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
 
     @GetMapping("/process")
     public ResponseEntity<List<Item>> processItems() {
